tree: move CleanTree to ignore
